+++ lib/plug_rest/router.ex
@@ -161,7 +161,13 @@ defmodule PlugRest.Router do
 
         handler_state = Keyword.get(options, :state)
 
-        PlugRest.Resource.upgrade(conn, unquote(handler), handler_state)
+        case function_exported?(unquote(handler), :call, 1) do
+          true ->
+            apply(unquote(handler), :call, [conn, options])
+          false ->
+            handler_state = Keyword.get(options, :state)
+            PlugRest.Resource.upgrade(conn, unquote(handler), handler_state)
+        end
       end
     end
   end