+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Backwards incompatible changes
   * Remove `read_path_params/1` (use `conn.params` instead)
   * Move private connection accessors from `Conn` to `Resource` module
+    and change the specs
 
 ## v0.6.2